+++ pkgs/maa-pipeline-manager/src/utils/helper.ts
@@ -34,6 +34,7 @@ export function extractTaskRef(r: TaskRefInfo): TaskName | null {
     r.type === 'task.anchor' ||
     r.type === 'task.reco' ||
     r.type === 'task.color_filter' ||
+    r.type === 'task.custom' ||
     r.type === 'task.entry'
   ) {
     return r.target