+++ pkgs/maa-tools/src/pm.ts
@@ -1,0 +1,1 @@
+export * from '@nekosu/maa-pipeline-manager'