+++ pkgs/maa-tools/src/types/config.ts
@@ -1,5 +1,5 @@
 import type { LocaleType } from '@nekosu/maa-locale'
-import type { DiagnosticType } from '@nekosu/maa-pipeline-manager'
+import type { DiagnosticType, ParserConfig } from '@nekosu/maa-pipeline-manager'
 import type { NpmRegistryType } from '@nekosu/maa-version-manager'
 
 export type BaseConfig = {
@@ -18,6 +18,8 @@ export type BaseConfig = {
   // color?: 'auto' | 'enable' | 'disable'
 
   interfacePath: string
+
+  parser?: ParserConfig
 }
 
 export type CheckConfig = {