+++ pkgs/maa-pipeline-manager/CHANGELOG.md
@@ -1,3 +1,7 @@
+- 1.0.8
+
+Support parser for custom reco & act
+
 - 1.0.7
 
 Fix tasks' pipeline_override missing